add support for multiple trackers based on compile-time options
Add an instance variable Tracker.config where each field will be used
to define a C flag at tracker compile time.

Each track kernel is cached with the corresponding config, so that
when the user changes the config to a previously seen one, the tracker
will not be recompiled again.

The above changes imply that trackers will be built on-demand, instead
of in the constructor. As a result the `skip` option for `track_kernel`
has been removed from the non-collective tracker.

Signed-off-by: Szymon Łopaciuk <szymon@lopaciuk.eu>