generate and reuse compiled track kernels
The track kernels can be precompiled using `xtrack.prebuild_kernels:
precompile_kernels`. Additionally the module contains the descriptions
of the kernels to be generated. When building the tracker, in
`xtrack.Tracker._build_tracker` an additional parameter
`built_ffi_module_name` can be passed, which will try to load the
kernel from the module of the given name, and failing that will
compile the kernel and save it as that module.

A new value for the parameter `compile` in `Tracker._build_tracker`
is introduced: it can now be either bool or 'force'. See the
docstring of the method for more details.

Signed-off-by: Szymon Łopaciuk <szymon@lopaciuk.eu>